Trade price is an amount

+++ crates/chia-sdk-types/src/condition.rs
@@ -171,7 +171,7 @@ conditions! {
         TransferNft as Default {
             opcode: i8 if -10,
             did_id: Option<Bytes32>,
-            trade_prices: Vec<(u16, Bytes32)>,
+            trade_prices: Vec<(u64, Bytes32)>,
             did_inner_puzzle_hash: Option<Bytes32>,
         },
         RunCatTail<P, S> as Copy {
